selinux-policy-2.eclass: support MODS being a bash array
Also update the various help comments to reflect that MODS is plural
and supports multiple modules listed in it.

If a non-existent module is listed in MODS, die; we're missing what is
in effect the source for our policy module, and cannot proceed. As MODS
contains the source for the policy package, we cannot afford to ignore
a non-existent module and proceed graciously.

If we're operating on the strict module store, COMMAND may not actually
contain any modules to install; the unconfined module is skipped over.
Hence, keep track of the number of modules being installed and return
early if there is nothing to install such that semodule does not error.

Signed-off-by: Rahul Sandhu <nvraxn@gmail.com>